+++ src/context/theme.ts
@@ -1,0 +1,12 @@
+import {MD3LightTheme as DefaultTheme} from 'react-native-paper';
+
+const theme = {
+  ...DefaultTheme,
+  colors: {
+    ...DefaultTheme.colors,
+    primary: 'tomato',
+    secondary: 'yellow',
+  },
+};
+
+export default theme;
